fix - based on comments from feedbacks in PR
