+++ forloop_modules/function_handlers/input_waiter.py
@@ -1,6 +1,7 @@
 # input_waiter.py
 from forloop_modules.globals.variable_handler import LocalVariable, variable_handler
 from forloop_modules.queries.node_context_requests_backend import post_console_log
+import forloop_modules.flog as flog
 
 _pending: dict[str, str] = {}
 _prompts: dict[str, str] = {} 
@@ -14,16 +15,12 @@ def resolve_input(node_uid: str, value: str):
     var_name = _pending.pop(node_uid, None)
     prompt_text = _prompts.pop(node_uid, None) 
 
-    if not var_name:
-        return
-
-    if var_name in variable_handler.variables:
-        variable_handler.variables[var_name] = LocalVariable(uid=0, name=var_name, value=value, is_result=False)
-    else:
+    if var_name is not None:
+        # Use the proper variable handler method for both new and existing variables
         variable_handler.new_variable(var_name, value)
 
-    try:
-        msg = f"{prompt_text} = {value}" if prompt_text else f"{var_name} = {value}"
-        post_console_log(message=msg, type="input")
-    except Exception:
-        pass
+        try:
+            msg = f"{prompt_text} = {value}" if prompt_text else f"{var_name} = {value}"
+            post_console_log(message=msg, type="input")
+        except Exception as e:
+            flog.debug(f"Failed to post console log: {e}")
